+++ map.c
@@ -10,7 +10,7 @@ struct _mapchip {
 	void *value;
 };
 
-#define GARBAGE_INT 0xCCCCCCCC
+#define GARBAGE_INT ((int)0xCCCCCCCC)
 
 /* it overrides; it does not overwrite. */
 void insert(struct map *ptr, const char *key, void *value)
